[Refactor] [9/N] to simplify the vLLM openai translations  serving architecture
