+++ vllm/entrypoints/openai/engine/serving.py
@@ -50,9 +50,6 @@ from vllm.entrypoints.openai.engine.protocol import (
     ErrorResponse,
     FunctionCall,
     FunctionDefinition,
-    TranscriptionRequest,
-    TranscriptionResponse,
-    TranslationRequest,
     VLLMValidationError,
 )
 from vllm.entrypoints.openai.responses.protocol import (
@@ -60,6 +57,11 @@ from vllm.entrypoints.openai.responses.protocol import (
     ResponsesRequest,
 )
 from vllm.entrypoints.openai.serving_models import OpenAIServingModels
+from vllm.entrypoints.openai.translations.protocol import (
+    TranscriptionRequest,
+    TranscriptionResponse,
+    TranslationRequest,
+)
 from vllm.entrypoints.pooling.classify.protocol import (
     ClassificationChatRequest,
     ClassificationCompletionRequest,
